chore(instance): extract server delete command from custom_server.go
